+++ src/sep.ira/calculator/NumericalRange.cs
@@ -6,11 +6,11 @@ public class NumericalRange
 {
     /// <summary>   Gets or sets the minimum. </summary>
     /// <value> The minimum value. </value>
-    public double Minimum { get; set; }
+    public decimal Minimum { get; set; }
 
     /// <summary>   Gets or sets the maximum. </summary>
     /// <value> The maximum value. </value>
-    public double Maximum { get; set; }
+    public decimal Maximum { get; set; }
 
     /// <summary>   Constructor. </summary>
     /// <remarks>   2026-06-15. </remarks>
@@ -18,7 +18,7 @@ public class NumericalRange
     ///                                         illegal values. </exception>
     /// <param name="minimum">  The minimum value. </param>
     /// <param name="maximum">  The maximum value. </param>
-    public NumericalRange( double minimum, double maximum )
+    public NumericalRange( decimal minimum, decimal maximum )
     {
         if ( minimum > maximum )
             throw new ArgumentException( "Minimum cannot be greater than maximum." );
@@ -29,7 +29,7 @@ public class NumericalRange
     /// <remarks>   2026-06-15. </remarks>
     /// <param name="value">    The value to check. </param>
     /// <returns>   True if the range contains the value; otherwise, false. </returns>
-    public bool Contains( double value )
+    public bool Contains( decimal value )
     {
         return value >= this.Minimum && value <= this.Maximum;
     }
